habeascorpus_tokens/lucene-3.6.2/contrib/analyzers/common/src/test/org/apache/lucene/analysis/el/GreekAnalyzerTest.java
package	TokenNamepackage	
org	TokenNameIdentifier	 org
.	TokenNameDOT	
apache	TokenNameIdentifier	 apache
.	TokenNameDOT	
lucene	TokenNameIdentifier	 lucene
.	TokenNameDOT	
analysis	TokenNameIdentifier	 analysis
.	TokenNameDOT	
el	TokenNameIdentifier	 el
;	TokenNameSEMICOLON	
/** * Copyright 2005 The Apache Software Foundation * * Licensed under the Apache License, Version 2.0 (the "License"); * you may not use this file except in compliance with the License. * You may obtain a copy of the License at * * http://www.apache.org/licenses/LICENSE-2.0 * * Unless required by applicable law or agreed to in writing, software * distributed under the License is distributed on an "AS IS" BASIS, * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. * See the License for the specific language governing permissions and * limitations under the License. */	TokenNameCOMMENT_JAVADOC	 Copyright 2005 The Apache Software Foundation * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at * http://www.apache.org/licenses/LICENSE-2.0 * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License. 
import	TokenNameimport	
org	TokenNameIdentifier	 org
.	TokenNameDOT	
apache	TokenNameIdentifier	 apache
.	TokenNameDOT	
lucene	TokenNameIdentifier	 lucene
.	TokenNameDOT	
analysis	TokenNameIdentifier	 analysis
.	TokenNameDOT	
BaseTokenStreamTestCase	TokenNameIdentifier	 Base Token Stream Test Case
;	TokenNameSEMICOLON	
import	TokenNameimport	
org	TokenNameIdentifier	 org
.	TokenNameDOT	
apache	TokenNameIdentifier	 apache
.	TokenNameDOT	
lucene	TokenNameIdentifier	 lucene
.	TokenNameDOT	
analysis	TokenNameIdentifier	 analysis
.	TokenNameDOT	
Analyzer	TokenNameIdentifier	 Analyzer
;	TokenNameSEMICOLON	
import	TokenNameimport	
org	TokenNameIdentifier	 org
.	TokenNameDOT	
apache	TokenNameIdentifier	 apache
.	TokenNameDOT	
lucene	TokenNameIdentifier	 lucene
.	TokenNameDOT	
util	TokenNameIdentifier	 util
.	TokenNameDOT	
Version	TokenNameIdentifier	 Version
;	TokenNameSEMICOLON	
/** * A unit test class for verifying the correct operation of the GreekAnalyzer. * */	TokenNameCOMMENT_JAVADOC	 A unit test class for verifying the correct operation of the GreekAnalyzer. 
public	TokenNamepublic	
class	TokenNameclass	
GreekAnalyzerTest	TokenNameIdentifier	 Greek Analyzer Test
extends	TokenNameextends	
BaseTokenStreamTestCase	TokenNameIdentifier	 Base Token Stream Test Case
{	TokenNameLBRACE	
/** * Test the analysis of various greek strings. * * @throws Exception in case an error occurs */	TokenNameCOMMENT_JAVADOC	 Test the analysis of various greek strings. * @throws Exception in case an error occurs 
public	TokenNamepublic	
void	TokenNamevoid	
testAnalyzer	TokenNameIdentifier	 test Analyzer
(	TokenNameLPAREN	
)	TokenNameRPAREN	
throws	TokenNamethrows	
Exception	TokenNameIdentifier	 Exception
{	TokenNameLBRACE	
Analyzer	TokenNameIdentifier	 Analyzer
a	TokenNameIdentifier	 a
=	TokenNameEQUAL	
new	TokenNamenew	
GreekAnalyzer	TokenNameIdentifier	 Greek Analyzer
(	TokenNameLPAREN	
TEST_VERSION_CURRENT	TokenNameIdentifier	 TEST  VERSION  CURRENT
)	TokenNameRPAREN	
;	TokenNameSEMICOLON	
// Verify the correct analysis of capitals and small accented letters, and 	TokenNameCOMMENT_LINE	Verify the correct analysis of capitals and small accented letters, and 
// stemming 	TokenNameCOMMENT_LINE	stemming 
assertAnalyzesTo	TokenNameIdentifier	 assert Analyzes To
(	TokenNameLPAREN	
a	TokenNameIdentifier	 a
,	TokenNameCOMMA	
"Μία εξαιρετικά καλή και πλούσια σειρά χαρακτήρων της Ελληνικής γλώσσας"	TokenNameStringLiteral	Μία εξαιρετικά καλή και πλούσια σειρά χαρακτήρων της Ελληνικής γλώσσας
,	TokenNameCOMMA	
new	TokenNamenew	
String	TokenNameIdentifier	 String
[	TokenNameLBRACKET	
]	TokenNameRBRACKET	
{	TokenNameLBRACE	
"μια"	TokenNameStringLiteral	μια
,	TokenNameCOMMA	
"εξαιρετ"	TokenNameStringLiteral	εξαιρετ
,	TokenNameCOMMA	
"καλ"	TokenNameStringLiteral	καλ
,	TokenNameCOMMA	
"πλουσ"	TokenNameStringLiteral	πλουσ
,	TokenNameCOMMA	
"σειρ"	TokenNameStringLiteral	σειρ
,	TokenNameCOMMA	
"χαρακτηρ"	TokenNameStringLiteral	χαρακτηρ
,	TokenNameCOMMA	
"ελληνικ"	TokenNameStringLiteral	ελληνικ
,	TokenNameCOMMA	
"γλωσσ"	TokenNameStringLiteral	γλωσσ
}	TokenNameRBRACE	
)	TokenNameRPAREN	
;	TokenNameSEMICOLON	
// Verify the correct analysis of small letters with diaeresis and the elimination 	TokenNameCOMMENT_LINE	Verify the correct analysis of small letters with diaeresis and the elimination 
// of punctuation marks 	TokenNameCOMMENT_LINE	of punctuation marks 
assertAnalyzesTo	TokenNameIdentifier	 assert Analyzes To
(	TokenNameLPAREN	
a	TokenNameIdentifier	 a
,	TokenNameCOMMA	
"Προϊόντα (και) [πολλαπλές] - ΑΝΑΓΚΕΣ"	TokenNameStringLiteral	Προϊόντα (και) [πολλαπλές] - ΑΝΑΓΚΕΣ
,	TokenNameCOMMA	
new	TokenNamenew	
String	TokenNameIdentifier	 String
[	TokenNameLBRACKET	
]	TokenNameRBRACKET	
{	TokenNameLBRACE	
"προιοντ"	TokenNameStringLiteral	προιοντ
,	TokenNameCOMMA	
"πολλαπλ"	TokenNameStringLiteral	πολλαπλ
,	TokenNameCOMMA	
"αναγκ"	TokenNameStringLiteral	αναγκ
}	TokenNameRBRACE	
)	TokenNameRPAREN	
;	TokenNameSEMICOLON	
// Verify the correct analysis of capital accented letters and capital letters with diaeresis, 	TokenNameCOMMENT_LINE	Verify the correct analysis of capital accented letters and capital letters with diaeresis, 
// as well as the elimination of stop words 	TokenNameCOMMENT_LINE	as well as the elimination of stop words 
assertAnalyzesTo	TokenNameIdentifier	 assert Analyzes To
(	TokenNameLPAREN	
a	TokenNameIdentifier	 a
,	TokenNameCOMMA	
"ΠΡΟΫΠΟΘΕΣΕΙΣ Άψογος, ο μεστός και οι άλλοι"	TokenNameStringLiteral	ΠΡΟΫΠΟΘΕΣΕΙΣ Άψογος, ο μεστός και οι άλλοι
,	TokenNameCOMMA	
new	TokenNamenew	
String	TokenNameIdentifier	 String
[	TokenNameLBRACKET	
]	TokenNameRBRACKET	
{	TokenNameLBRACE	
"προυποθεσ"	TokenNameStringLiteral	προυποθεσ
,	TokenNameCOMMA	
"αψογ"	TokenNameStringLiteral	αψογ
,	TokenNameCOMMA	
"μεστ"	TokenNameStringLiteral	μεστ
,	TokenNameCOMMA	
"αλλ"	TokenNameStringLiteral	αλλ
}	TokenNameRBRACE	
)	TokenNameRPAREN	
;	TokenNameSEMICOLON	
}	TokenNameRBRACE	
/** * Test the analysis of various greek strings. * * @throws Exception in case an error occurs * @deprecated Remove this test when support for 3.0 is no longer needed */	TokenNameCOMMENT_JAVADOC	 Test the analysis of various greek strings. * @throws Exception in case an error occurs @deprecated Remove this test when support for 3.0 is no longer needed 
@	TokenNameAT	
Deprecated	TokenNameIdentifier	 Deprecated
public	TokenNamepublic	
void	TokenNamevoid	
testAnalyzerBWCompat	TokenNameIdentifier	 test Analyzer BW Compat
(	TokenNameLPAREN	
)	TokenNameRPAREN	
throws	TokenNamethrows	
Exception	TokenNameIdentifier	 Exception
{	TokenNameLBRACE	
Analyzer	TokenNameIdentifier	 Analyzer
a	TokenNameIdentifier	 a
=	TokenNameEQUAL	
new	TokenNamenew	
GreekAnalyzer	TokenNameIdentifier	 Greek Analyzer
(	TokenNameLPAREN	
Version	TokenNameIdentifier	 Version
.	TokenNameDOT	
LUCENE_30	TokenNameIdentifier	 LUCENE 30
)	TokenNameRPAREN	
;	TokenNameSEMICOLON	
// Verify the correct analysis of capitals and small accented letters 	TokenNameCOMMENT_LINE	Verify the correct analysis of capitals and small accented letters 
assertAnalyzesTo	TokenNameIdentifier	 assert Analyzes To
(	TokenNameLPAREN	
a	TokenNameIdentifier	 a
,	TokenNameCOMMA	
"Μία εξαιρετικά καλή και πλούσια σειρά χαρακτήρων της Ελληνικής γλώσσας"	TokenNameStringLiteral	Μία εξαιρετικά καλή και πλούσια σειρά χαρακτήρων της Ελληνικής γλώσσας
,	TokenNameCOMMA	
new	TokenNamenew	
String	TokenNameIdentifier	 String
[	TokenNameLBRACKET	
]	TokenNameRBRACKET	
{	TokenNameLBRACE	
"μια"	TokenNameStringLiteral	μια
,	TokenNameCOMMA	
"εξαιρετικα"	TokenNameStringLiteral	εξαιρετικα
,	TokenNameCOMMA	
"καλη"	TokenNameStringLiteral	καλη
,	TokenNameCOMMA	
"πλουσια"	TokenNameStringLiteral	πλουσια
,	TokenNameCOMMA	
"σειρα"	TokenNameStringLiteral	σειρα
,	TokenNameCOMMA	
"χαρακτηρων"	TokenNameStringLiteral	χαρακτηρων
,	TokenNameCOMMA	
"ελληνικησ"	TokenNameStringLiteral	ελληνικησ
,	TokenNameCOMMA	
"γλωσσασ"	TokenNameStringLiteral	γλωσσασ
}	TokenNameRBRACE	
)	TokenNameRPAREN	
;	TokenNameSEMICOLON	
// Verify the correct analysis of small letters with diaeresis and the elimination 	TokenNameCOMMENT_LINE	Verify the correct analysis of small letters with diaeresis and the elimination 
// of punctuation marks 	TokenNameCOMMENT_LINE	of punctuation marks 
assertAnalyzesTo	TokenNameIdentifier	 assert Analyzes To
(	TokenNameLPAREN	
a	TokenNameIdentifier	 a
,	TokenNameCOMMA	
"Προϊόντα (και) [πολλαπλές] - ΑΝΑΓΚΕΣ"	TokenNameStringLiteral	Προϊόντα (και) [πολλαπλές] - ΑΝΑΓΚΕΣ
,	TokenNameCOMMA	
new	TokenNamenew	
String	TokenNameIdentifier	 String
[	TokenNameLBRACKET	
]	TokenNameRBRACKET	
{	TokenNameLBRACE	
"προιοντα"	TokenNameStringLiteral	προιοντα
,	TokenNameCOMMA	
"πολλαπλεσ"	TokenNameStringLiteral	πολλαπλεσ
,	TokenNameCOMMA	
"αναγκεσ"	TokenNameStringLiteral	αναγκεσ
}	TokenNameRBRACE	
)	TokenNameRPAREN	
;	TokenNameSEMICOLON	
// Verify the correct analysis of capital accented letters and capital letters with diaeresis, 	TokenNameCOMMENT_LINE	Verify the correct analysis of capital accented letters and capital letters with diaeresis, 
// as well as the elimination of stop words 	TokenNameCOMMENT_LINE	as well as the elimination of stop words 
assertAnalyzesTo	TokenNameIdentifier	 assert Analyzes To
(	TokenNameLPAREN	
a	TokenNameIdentifier	 a
,	TokenNameCOMMA	
"ΠΡΟΫΠΟΘΕΣΕΙΣ Άψογος, ο μεστός και οι άλλοι"	TokenNameStringLiteral	ΠΡΟΫΠΟΘΕΣΕΙΣ Άψογος, ο μεστός και οι άλλοι
,	TokenNameCOMMA	
new	TokenNamenew	
String	TokenNameIdentifier	 String
[	TokenNameLBRACKET	
]	TokenNameRBRACKET	
{	TokenNameLBRACE	
"προυποθεσεισ"	TokenNameStringLiteral	προυποθεσεισ
,	TokenNameCOMMA	
"αψογοσ"	TokenNameStringLiteral	αψογοσ
,	TokenNameCOMMA	
"μεστοσ"	TokenNameStringLiteral	μεστοσ
,	TokenNameCOMMA	
"αλλοι"	TokenNameStringLiteral	αλλοι
}	TokenNameRBRACE	
)	TokenNameRPAREN	
;	TokenNameSEMICOLON	
}	TokenNameRBRACE	
public	TokenNamepublic	
void	TokenNamevoid	
testReusableTokenStream	TokenNameIdentifier	 test Reusable Token Stream
(	TokenNameLPAREN	
)	TokenNameRPAREN	
throws	TokenNamethrows	
Exception	TokenNameIdentifier	 Exception
{	TokenNameLBRACE	
Analyzer	TokenNameIdentifier	 Analyzer
a	TokenNameIdentifier	 a
=	TokenNameEQUAL	
new	TokenNamenew	
GreekAnalyzer	TokenNameIdentifier	 Greek Analyzer
(	TokenNameLPAREN	
TEST_VERSION_CURRENT	TokenNameIdentifier	 TEST  VERSION  CURRENT
)	TokenNameRPAREN	
;	TokenNameSEMICOLON	
// Verify the correct analysis of capitals and small accented letters, and 	TokenNameCOMMENT_LINE	Verify the correct analysis of capitals and small accented letters, and 
// stemming 	TokenNameCOMMENT_LINE	stemming 
assertAnalyzesToReuse	TokenNameIdentifier	 assert Analyzes To Reuse
(	TokenNameLPAREN	
a	TokenNameIdentifier	 a
,	TokenNameCOMMA	
"Μία εξαιρετικά καλή και πλούσια σειρά χαρακτήρων της Ελληνικής γλώσσας"	TokenNameStringLiteral	Μία εξαιρετικά καλή και πλούσια σειρά χαρακτήρων της Ελληνικής γλώσσας
,	TokenNameCOMMA	
new	TokenNamenew	
String	TokenNameIdentifier	 String
[	TokenNameLBRACKET	
]	TokenNameRBRACKET	
{	TokenNameLBRACE	
"μια"	TokenNameStringLiteral	μια
,	TokenNameCOMMA	
"εξαιρετ"	TokenNameStringLiteral	εξαιρετ
,	TokenNameCOMMA	
"καλ"	TokenNameStringLiteral	καλ
,	TokenNameCOMMA	
"πλουσ"	TokenNameStringLiteral	πλουσ
,	TokenNameCOMMA	
"σειρ"	TokenNameStringLiteral	σειρ
,	TokenNameCOMMA	
"χαρακτηρ"	TokenNameStringLiteral	χαρακτηρ
,	TokenNameCOMMA	
"ελληνικ"	TokenNameStringLiteral	ελληνικ
,	TokenNameCOMMA	
"γλωσσ"	TokenNameStringLiteral	γλωσσ
}	TokenNameRBRACE	
)	TokenNameRPAREN	
;	TokenNameSEMICOLON	
// Verify the correct analysis of small letters with diaeresis and the elimination 	TokenNameCOMMENT_LINE	Verify the correct analysis of small letters with diaeresis and the elimination 
// of punctuation marks 	TokenNameCOMMENT_LINE	of punctuation marks 
assertAnalyzesToReuse	TokenNameIdentifier	 assert Analyzes To Reuse
(	TokenNameLPAREN	
a	TokenNameIdentifier	 a
,	TokenNameCOMMA	
"Προϊόντα (και) [πολλαπλές] - ΑΝΑΓΚΕΣ"	TokenNameStringLiteral	Προϊόντα (και) [πολλαπλές] - ΑΝΑΓΚΕΣ
,	TokenNameCOMMA	
new	TokenNamenew	
String	TokenNameIdentifier	 String
[	TokenNameLBRACKET	
]	TokenNameRBRACKET	
{	TokenNameLBRACE	
"προιοντ"	TokenNameStringLiteral	προιοντ
,	TokenNameCOMMA	
"πολλαπλ"	TokenNameStringLiteral	πολλαπλ
,	TokenNameCOMMA	
"αναγκ"	TokenNameStringLiteral	αναγκ
}	TokenNameRBRACE	
)	TokenNameRPAREN	
;	TokenNameSEMICOLON	
// Verify the correct analysis of capital accented letters and capital letters with diaeresis, 	TokenNameCOMMENT_LINE	Verify the correct analysis of capital accented letters and capital letters with diaeresis, 
// as well as the elimination of stop words 	TokenNameCOMMENT_LINE	as well as the elimination of stop words 
assertAnalyzesToReuse	TokenNameIdentifier	 assert Analyzes To Reuse
(	TokenNameLPAREN	
a	TokenNameIdentifier	 a
,	TokenNameCOMMA	
"ΠΡΟΫΠΟΘΕΣΕΙΣ Άψογος, ο μεστός και οι άλλοι"	TokenNameStringLiteral	ΠΡΟΫΠΟΘΕΣΕΙΣ Άψογος, ο μεστός και οι άλλοι
,	TokenNameCOMMA	
new	TokenNamenew	
String	TokenNameIdentifier	 String
[	TokenNameLBRACKET	
]	TokenNameRBRACKET	
{	TokenNameLBRACE	
"προυποθεσ"	TokenNameStringLiteral	προυποθεσ
,	TokenNameCOMMA	
"αψογ"	TokenNameStringLiteral	αψογ
,	TokenNameCOMMA	
"μεστ"	TokenNameStringLiteral	μεστ
,	TokenNameCOMMA	
"αλλ"	TokenNameStringLiteral	αλλ
}	TokenNameRBRACE	
)	TokenNameRPAREN	
;	TokenNameSEMICOLON	
}	TokenNameRBRACE	
/** * Greek Analyzer didn't call standardFilter, so no normalization of acronyms. * check that this is preserved. * @deprecated remove this test in Lucene 4.0 */	TokenNameCOMMENT_JAVADOC	 Greek Analyzer didn't call standardFilter, so no normalization of acronyms. check that this is preserved. @deprecated remove this test in Lucene 4.0 
@	TokenNameAT	
Deprecated	TokenNameIdentifier	 Deprecated
public	TokenNamepublic	
void	TokenNamevoid	
testAcronymBWCompat	TokenNameIdentifier	 test Acronym BW Compat
(	TokenNameLPAREN	
)	TokenNameRPAREN	
throws	TokenNamethrows	
Exception	TokenNameIdentifier	 Exception
{	TokenNameLBRACE	
Analyzer	TokenNameIdentifier	 Analyzer
a	TokenNameIdentifier	 a
=	TokenNameEQUAL	
new	TokenNamenew	
GreekAnalyzer	TokenNameIdentifier	 Greek Analyzer
(	TokenNameLPAREN	
Version	TokenNameIdentifier	 Version
.	TokenNameDOT	
LUCENE_30	TokenNameIdentifier	 LUCENE 30
)	TokenNameRPAREN	
;	TokenNameSEMICOLON	
assertAnalyzesTo	TokenNameIdentifier	 assert Analyzes To
(	TokenNameLPAREN	
a	TokenNameIdentifier	 a
,	TokenNameCOMMA	
"Α.Π.Τ."	TokenNameStringLiteral	Α.Π.Τ.
,	TokenNameCOMMA	
new	TokenNamenew	
String	TokenNameIdentifier	 String
[	TokenNameLBRACKET	
]	TokenNameRBRACKET	
{	TokenNameLBRACE	
"α.π.τ."	TokenNameStringLiteral	α.π.τ.
}	TokenNameRBRACE	
)	TokenNameRPAREN	
;	TokenNameSEMICOLON	
}	TokenNameRBRACE	
/** blast some random strings through the analyzer */	TokenNameCOMMENT_JAVADOC	 blast some random strings through the analyzer 
public	TokenNamepublic	
void	TokenNamevoid	
testRandomStrings	TokenNameIdentifier	 test Random Strings
(	TokenNameLPAREN	
)	TokenNameRPAREN	
throws	TokenNamethrows	
Exception	TokenNameIdentifier	 Exception
{	TokenNameLBRACE	
checkRandomData	TokenNameIdentifier	 check Random Data
(	TokenNameLPAREN	
random	TokenNameIdentifier	 random
,	TokenNameCOMMA	
new	TokenNamenew	
GreekAnalyzer	TokenNameIdentifier	 Greek Analyzer
(	TokenNameLPAREN	
TEST_VERSION_CURRENT	TokenNameIdentifier	 TEST  VERSION  CURRENT
)	TokenNameRPAREN	
,	TokenNameCOMMA	
10000	TokenNameIntegerLiteral	
*	TokenNameMULTIPLY	
RANDOM_MULTIPLIER	TokenNameIdentifier	 RANDOM  MULTIPLIER
)	TokenNameRPAREN	
;	TokenNameSEMICOLON	
}	TokenNameRBRACE	
}	TokenNameRBRACE	
